requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr # Apache-2.0

eventlet # MIT
netaddr # BSD
six # MIT
stevedore # Apache-2.0
Django
django-compressor
django-debug-toolbar
django-extensions
ipython
psycopg2
six
wsgiref
pyyaml
MySQL-python
alembic
